(relay_agent) Fix guacamole_baseurl
For relays, we used the FQDN of the relay server for the
guacamole_baseurl. But we cannot as we contact first the principal.

Now we use the pulsemain FQDN.

In case where the relays are not accessible outside ( siveo.lan for
example ) we cannot contact it but apache does the call.