(ansible pulse_relay) Clear glpi cache
When we create an entity by the API, there is a bug in GLPI and the
Entities are not usable. We need to use the command:

php /usr/share/glpi/bin/console cache:clear to make it OK.

Delegate this to the main server.

(cherry picked from commit 64fc32bf8d41f7ba5d51dc1a31f367a67e3f21a9)